Fetching woody plants by filter implemented

+++ src/TreeLoc.Api/Models/WoodyPlantListModel.cs
@@ -5,6 +5,6 @@ namespace TreeLoc.Api.Models
   public class WoodyPlantListModel
   {
     public List<WoodyPlantPreviewModel> WoodyPlants { get; set; } = default!;
-    public int TotalCount { get; set; }
+    public long TotalCount { get; set; }
   }
 }
